OpenTracing.Propagation

Description

Types and functions for serializing and deserializing SpanContexts across process boundaries.

One of the big motiviating use cases for propagation is for tracing distributed executions through RPC calls.

Documentation

type TextMap = HashMap Text Text

type Headers = [Header]

type Propagation carriers = Rec Carrier carriers

A Propagation contains the different ways that a SpanContext can be serialized and deserialized. For example Propagation '[TextMap, Headers] indicates support for serializing to Header or to TextMap.

Since: 0.1.0.0

class HasPropagation a p | a -> p where

A typeclass for application environments that contain a Propagation.

Since: 0.1.0.0

Methods

propagation :: Getting r a (Propagation p)

newtype Carrier a

`Carrier a` is a way to convert a SpanContext into or from an a.

Since: 0.1.0.0

Constructors

Carrier

Fields

fromCarrier :: Prism' a SpanContext

type HasCarrier c cs = c ∈ cs

type HasCarriers cs ds = cs ⊆ ds

carrier

Arguments

:: (HasCarrier c cs, HasPropagation r cs)

=> proxy c
Proxy for the carrier type c.

-> r
The application context

-> Prism' c SpanContext

Retrieve a (de)serialization lens from the application context for format c.

Since: 0.1.0.0

inject :: forall c r p. (HasCarrier c p, HasPropagation r p) => r -> SpanContext -> c

Serialize a SpanContext into the format c using a serializer from the application context.

Since: 0.1.0.0

extract :: forall c r p. (HasCarrier c p, HasPropagation r p) => r -> c -> Maybe SpanContext

Attempt to deserialize a SpanContext from the format c using a deserializer from the application context

Since: 0.1.0.0

otPropagation :: Propagation '[TextMap, Headers]

A propagation using an "ot" prefix. No parent span id is propagated in OT.

b3Propagation :: Propagation '[TextMap, Headers]

A propagation using an "x-b3" prefix for use with Zipkin.

_OTTextMap :: Prism' TextMap SpanContext

_OTHeaders :: Prism' Headers SpanContext

_B3TextMap :: Prism' TextMap SpanContext

_B3Headers :: Prism' Headers SpanContext

_HeadersTextMap :: Iso' Headers TextMap

Convert between a TextMap and Headers

Header field values are URL-encoded when converting from TextMap to Headers, and URL-decoded when converting the other way.

Note: validity of header fields is not checked (RFC 7230, 3.2.4)

Re-exports from Vinyl

data Rec (a :: u -> Type) (b :: [u]) :: forall u. (u -> Type) -> [u] -> Type where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type) (b :: [u]). Rec a ([] :: [u])

(:&) :: forall u (a :: u -> Type) (b :: [u]) (r :: u) (rs :: [u]). !(a r) -> !(Rec a rs) -> Rec a (r ': rs) infixr 7

rappend :: Rec f as -> Rec f bs -> Rec f (as ++ bs)

Two records may be pasted together.

(<+>) :: Rec f as -> Rec f bs -> Rec f (as ++ bs) infixr 5

A shorthand for rappend.

rcast :: (RecSubset record rs ss is, RecSubsetFCtx record f) => record f ss -> record f rs

Takes a larger record to a smaller one by forgetting fields. This is rcastC with the type arguments reordered for more convenient usage with TypeApplications.