OpenTracing.Reporting.Batch

Description

This module provides a trace reporter that groups recorded spans into batches before sending them to their destination in bulk.

Documentation

data BatchOptions

Options available to construct a batch reporter. Default options are available with batchOptions

batchOptions :: ([FinishedSpan] -> IO ()) -> BatchOptions

Default batch options which can be overridden via lenses.

boptBatchSize :: Lens' BatchOptions Word16

boptTimeoutSec :: Lens' BatchOptions Word

boptReporter :: Lens' BatchOptions ([FinishedSpan] -> IO ())

boptErrorLog :: Lens' BatchOptions (Builder -> IO ())

defaultErrorLog :: Builder -> IO ()

An error logging function which prints to stderr.

data BatchEnv

The environment of a batch reporter.

newBatchEnv :: BatchOptions -> IO BatchEnv

Create a new batch environment

closeBatchEnv :: BatchEnv -> IO ()

Close a batch reporter, stop consuming any new spans. Any spans in the queue will be drained.

batchReporter :: MonadIO m => BatchEnv -> FinishedSpan -> m ()

An implementation of tracerReport that batches the finished spans for transimission to their destination.