OpenTracing.Reporting.Pure

Description

Reporters with no external components.

Documentation

noReporter :: MonadIO m => FinishedSpan -> m ()

A null reporter which ignores anything it's given.

memReporter :: MonadIO m => Mem -> FinishedSpan -> m ()

A reporter which stores the finished spans in memory where they wait to be consumed.

newMem :: IO Mem

Construct a new memReporter environment that can store an unbounded seequence of FinishedSpans.

newBoundedMem :: Word32 -> IO Mem

Construct a new memReporter environment that stores a bounded sequence of FinishedSpans

memPeek :: Mem -> IO [FinishedSpan]

View the FinishedSpans in a memReporter without removing them.

memTake :: Mem -> IO [FinishedSpan]

View and remove the FinishedSpans in a memReporter.