OpenTracing.Span

Description

Data types and functions for manipulating spans

Documentation

data SpanContext

A SpanContext is the data that uniquely identifies a span and the context in which it occurs. Spans occur in traces, which form complete pictures of a computation, potentially across multiple machines.

Since: 0.1.0.0

Constructors

SpanContext

Fields

ctxTraceID :: TraceID

A trace identifier. Trace ids are globally unique

ctxSpanID :: Word64

A span identifier. Span identifiers are unique to their trace.

ctxParentSpanID :: Maybe Word64

Spans without a parent are known as "root spans"

_ctxSampled :: Sampled

Whether or not this span is going to be reported.

_ctxBaggage :: HashMap Text Text

Baggage is arbitrary key:value pairs that cross process boundaries.

ctxSampled :: Lens' SpanContext Sampled

ctxBaggage :: Lens' SpanContext (HashMap Text Text)

data Span

Span is a span that has been started (but not finished). See the OpenTracing spec for more info

Since: 0.1.0.0

newSpan :: (MonadIO m, Foldable t) => SpanContext -> Text -> SpanRefs -> t Tag -> m Span

Create a new Span with the provided info. The created Span will have a start time equal to the system time when newSpan is called.

Since: 0.1.0.0

class HasSpanFields a

Minimal complete definition

spanContext, spanOperation, spanStart, spanTags, spanLogs

data ActiveSpan

A mutable Span that is currently being recorded.

Since: 0.1.0.0

mkActive :: Span -> IO ActiveSpan

Since: 0.1.0.0

modifyActiveSpan :: ActiveSpan -> (Span -> Span) -> IO ()

Since: 0.1.0.0

readActiveSpan :: ActiveSpan -> IO Span

Since: 0.1.0.0

addLogRecord :: ActiveSpan -> LogField -> IO ()

Log structured data to an ActiveSpan. More info in the OpenTracing spec

Since: 0.1.0.0

addLogRecord' :: ActiveSpan -> LogField -> [LogField] -> IO ()

setBaggageItem :: ActiveSpan -> Text -> Text -> IO ()

getBaggageItem :: ActiveSpan -> Text -> IO (Maybe Text)

data FinishedSpan

A span that has finished executing.

Since: 0.1.0.0

spanFinish :: MonadIO m => Span -> m FinishedSpan

Convert an unfinished Span into a FinishedSpan

Since: 0.1.0.0

spanContext :: HasSpanFields a => Lens' a SpanContext

spanOperation :: HasSpanFields a => Lens' a Text

spanStart :: HasSpanFields a => Lens' a UTCTime

spanTags :: HasSpanFields a => Lens' a Tags

spanRefs :: HasRefs s a => Lens' s a

spanLogs :: HasSpanFields a => Lens' a [LogRecord]

spanDuration :: Lens' FinishedSpan NominalDiffTime

data SpanOpts

SpanOpts is the metadata information about a span needed in order to start measuring a span. This is the information that application code will provide in order to indicate what a span is doing and how it related to other spans. More info in the OpenTracing spec

Since: 0.1.0.0

spanOpts :: Text -> SpanRefs -> SpanOpts

Create a new SpanOpts with the minimal amount of required information.

Since: 0.1.0.0

spanOptOperation :: Lens' SpanOpts Text

spanOptRefs :: Lens' SpanOpts SpanRefs

spanOptTags :: Lens' SpanOpts [Tag]

spanOptSampled :: Lens' SpanOpts (Maybe Sampled)

data Reference

A reference from one span to another. Spans can be related in two ways:

ChildOf indicates that the parent span is dependent on the child span in order to produce its own result.

FollowsFrom indicates that there is no dependence relation, perhaps the parent span spawned an asynchronous task.

More info in the OpenTracing spec

Since: 0.1.0.0

Constructors

ChildOf

Fields

refCtx :: SpanContext

FollowsFrom

Fields

refCtx :: SpanContext

findParent :: Foldable t => t Reference -> Maybe Reference

data SpanRefs

The different references that a span can hold to other spans.

Since: 0.1.0.0

refActiveParents :: Lens' SpanRefs [ActiveSpan]

refPredecessors :: Lens' SpanRefs [FinishedSpan]

refPropagated :: Lens' SpanRefs [Reference]

childOf :: ActiveSpan -> SpanRefs

Create a SpanRefs containing the single refrence to a parent span.

Since: 0.1.0.0

followsFrom :: FinishedSpan -> SpanRefs

Create a SpanRefs containing the single refrence to a predecessor span.

Since: 0.1.0.0

freezeRefs :: SpanRefs -> IO [Reference]

Convert SpanRefs (which may include the mutable ActiveSpans) into an immutable list of References

Since: 0.1.0.0

data Sampled

A datatype indicating whether a recorded span was sampled, i.e. whether or not it will be reported. Traces are often sampled in high volume environments to keep the amount of data generated manageable.

Since: 0.1.0.0

Constructors

NotSampled

Sampled

_IsSampled :: Iso' Bool Sampled

sampled :: HasSampled a => Lens' a Sampled

data Traced a

A wrapper for a value that was produced by a traced computation.

Since: 0.1.0.0

Constructors

Traced

Fields

tracedResult :: a

The raw value produced

tracedSpan :: ~FinishedSpan

The resulting span that was created