OpenTracing.Tags

Documentation

data Tags

Tags are structured data associated with a Span. They can give a more complete picture of what a Span is doing than the operation alone. Tags apply to the entire timerange of a Span. Use LogField for events that refer to particular timestamp.

type Tag = (Text, TagVal)

A Tag is a key:value pair

data TagVal

Constructors

BoolT Bool

StringT Text

IntT Int64

DoubleT Double

BinaryT ByteString

setTag :: Tag -> Tags -> Tags

getTag :: Text -> Tags -> Maybe TagVal

getTagReify :: Getting (First b) Tag b -> Text -> Tags -> Maybe b

Get a tag and attempt to convert it from a serialized format

Standard span tags.

Refer to the OpenTracing spec for more info.

pattern ComponentKey :: forall a. (Eq a, IsString a) => a

pattern DbInstanceKey :: forall a. (Eq a, IsString a) => a

pattern DbStatementKey :: forall a. (Eq a, IsString a) => a

pattern DbTypeKey :: forall a. (Eq a, IsString a) => a

pattern DbUserKey :: forall a. (Eq a, IsString a) => a

pattern ErrorKey :: forall a. (Eq a, IsString a) => a

pattern HttpMethodKey :: forall a. (Eq a, IsString a) => a

pattern HttpStatusCodeKey :: forall a. (Eq a, IsString a) => a

pattern HttpUrlKey :: forall a. (Eq a, IsString a) => a

pattern MessageBusDestinationKey :: forall a. (Eq a, IsString a) => a

pattern PeerAddressKey :: forall a. (Eq a, IsString a) => a

pattern PeerHostnameKey :: forall a. (Eq a, IsString a) => a

pattern PeerIPv4Key :: forall a. (Eq a, IsString a) => a

pattern PeerIPv6Key :: forall a. (Eq a, IsString a) => a

pattern PeerPortKey :: forall a. (Eq a, IsString a) => a

pattern PeerServiceKey :: forall a. (Eq a, IsString a) => a

pattern SamplingPriorityKey :: forall a. (Eq a, IsString a) => a

pattern SpanKindKey :: forall a. (Eq a, IsString a) => a

pattern Component :: Text -> Tag

pattern DbInstance :: Text -> Tag

pattern DbStatement :: Text -> Tag

pattern DbType :: Text -> Tag

pattern DbUser :: Text -> Tag

pattern Error :: Bool -> Tag

pattern HttpMethod :: Method -> Tag

pattern HttpStatusCode :: Status -> Tag

pattern HttpUrl :: Text -> Tag

pattern MessageBusDestination :: Text -> Tag

pattern PeerAddress :: Text -> Tag

pattern PeerHostname :: Text -> Tag

pattern PeerIPv4 :: IPv4 -> Tag

pattern PeerIPv6 :: IPv6 -> Tag

pattern PeerPort :: Port -> Tag

pattern PeerService :: Text -> Tag

pattern SamplingPriority :: Word8 -> Tag

pattern SpanKind :: SpanKinds -> Tag

_Component :: Prism' Tag Text

_DbInstance :: Prism' Tag Text

_DbStatement :: Prism' Tag Text

_DbType :: Prism' Tag Text

_DbUser :: Prism' Tag Text

_Error :: Prism' Tag Bool

_HttpMethod :: Prism' Tag Method

_HttpStatusCode :: Prism' Tag Status

_HttpUrl :: Prism' Tag Text

_MessageBusDestination :: Prism' Tag Text

_PeerAddress :: Prism' Tag Text

_PeerHostname :: Prism' Tag Text

_PeerIPv4 :: Prism' Tag IPv4

_PeerIPv6 :: Prism' Tag IPv6

_PeerPort :: Prism' Tag Port

_PeerService :: Prism' Tag Text

_SamplingPriority :: Prism' Tag Word8

_SpanKind :: Prism' Tag SpanKinds

data SpanKinds

Constructors

RPCClient

RPCServer

Producer

Consumer

spanKindLabel :: SpanKinds -> Text